moved BirthdayProfiles storage responsibilities into a ProfileStore class
